Collect framework components in separate namespace
