+++ src/BrowserPicker/ExceptionModel.cs
@@ -1,4 +1,5 @@
-﻿using JetBrains.Annotations;
+﻿using BrowserPicker.Framework;
+using JetBrains.Annotations;
 using System;
 
 namespace BrowserPicker
