+++ src/BrowserPicker/Framework/RegistryHelpers.cs
@@ -4,7 +4,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Runtime.CompilerServices;
 
-namespace BrowserPicker
+namespace BrowserPicker.Framework
 {
 	public static class RegistryHelpers
 	{
