+++ src/BrowserPicker/Framework/ViewModelBase.cs
@@ -1,4 +1,4 @@
-namespace BrowserPicker
+namespace BrowserPicker.Framework
 {
 	public abstract class ViewModelBase<T> : ModelBase where T : ModelBase
 	{
